+++ src/core.ts
@@ -8,12 +8,25 @@ enum CopyCommand {
     CopyAbsolutePathWithLine,
 }
 
+class Range {
+    start: number;
+    end: number;
+    constructor(start: number, end: number) {
+        this.start = start;
+        this.end = end;
+    }
+}
+
 function GetContent(command: CopyCommand, uri: Uri): string {
     var editor = window.activeTextEditor;
 
     var absolutePath: string;
     var currentLine: number;
+    var selectedLines: string;
+    var selectionRanges: Range[] = new Array<Range>();
+    var isSingleLine = true;
 
+    var lineInfo: string | number;
 
     if (!editor) {
         // 如果没有打开的编辑器，那么返回 uri 的对应路径
@@ -21,9 +34,15 @@ function GetContent(command: CopyCommand, uri: Uri): string {
         currentLine = 1;
     } else {
         absolutePath = editor.document.fileName;
+        isSingleLine = editor.selections.length === 1 && editor.selections[0].isSingleLine;
         currentLine = editor.selection.active.line + 1;
+        editor.selections.forEach(selection => {
+            selectionRanges.push(new Range(selection.start.line + 1, selection.end.line + 1));
+        });
     }
 
+
+    // use uri.fsPath in title context 
     if (absolutePath !== uri.fsPath) {
         absolutePath = uri.fsPath;
         currentLine = 1;
@@ -31,9 +50,24 @@ function GetContent(command: CopyCommand, uri: Uri): string {
 
     var relativePath = workspace.asRelativePath(absolutePath);
 
+    // there is no item selected, so set currentLine to zero
     var workspaceDir = workspace.getWorkspaceFolder(uri);
     if (workspaceDir && workspaceDir.uri.fsPath === uri.fsPath) {
         currentLine = 0;
+    }
+
+
+    if (isSingleLine) {
+        lineInfo = currentLine;
+    } else {
+        selectedLines = selectionRanges.map(range => {
+            if (range.start === range.end) {
+                return range.start;
+            }
+
+            return `${range.start}~${range.end}`;
+        }).join(", ");
+        lineInfo = selectedLines;
     }
 
 
@@ -43,9 +77,9 @@ function GetContent(command: CopyCommand, uri: Uri): string {
         case CopyCommand.CopyAbsolutePath:
             return absolutePath;
         case CopyCommand.CopyRelativePathWithLine:
-            return `${relativePath}:${currentLine}`;
+            return `${relativePath}:${lineInfo}`;
         case CopyCommand.CopyAbsolutePathWithLine:
-            return `${absolutePath}:${currentLine}`;
+            return `${absolutePath}:${lineInfo}`;
         default:
             return "not supported command";
     }