+++ src/extension.ts
@@ -2,11 +2,6 @@ import * as vscode from 'vscode';
 import { DoCopy } from './command';
 import { CopyCommand } from './core';
 
-// "qishan233.copy.relative.path",
-// "qishan233.copy.absolute.path",
-// "qishan233.copy.relative.path.line",
-// "qishan233.copy.absolute.path.line",
-
 export function activate(context: vscode.ExtensionContext) {
 	console.log('Congratulations, your extension "copy-path-with-line-number" is now active!');