+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to the "copy-path-with-line-number" extension will be docume
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.2] - 2024-07-28
 
-- Add the ability of copying path(relative path and absolute path) with line number
+### Added
+
+- Add the ability of copying multiple selections line number info in editor context.
+
+## [0.0.1] - 2024-07-27
+
+### Added
+
+- Add the ability of copying path(relative path and absolute path) with line number.